Beta version 0.2.0! / Версия Beta 0.2.0! - improved the trading system: -- added a field for entering the quantity of the resource being sold; -- the load is measured in integer units without floating point; ---------------------------------------------------------------- - улучшил систему торговли: -- добавил поле для ввода количества продаваемого ресурса; -- груз измеряется в целочисленных единицах без плавающей запятой;

+++ Assets/Scripts/Cargos/Cargo.cs
@@ -6,15 +6,15 @@ namespace PlanetsColony
 {
     public class Cargo : Resource
     {
-        public Cargo(Type type, float value) : base(type, value)
+        public Cargo(Type type, uint value) : base(type, value)
         {
             this.value = value;
             this._type = type;
         }
 
-        public Cargo(Type type, float minValue, float maxValue) : base(type, minValue, maxValue)
+        public Cargo(Type type, uint minValue, uint maxValue) : base(type, minValue, maxValue)
         {
-            this.value = Random.Range(minValue, maxValue);
+            this.value = (uint)Random.Range(minValue, maxValue);
             this._type = type;
         }
 
@@ -23,7 +23,7 @@ namespace PlanetsColony
             return 0f;
         }
 
-        public override float GetValue()
+        public override uint GetValue()
         {
             return value;
         }
