+++ Assets/Scripts/Cargos/CargoGenerator.cs
@@ -4,7 +4,7 @@ namespace PlanetsColony
 {
     public class CargoGenerator : MonoBehaviour
     {
-        public Cargo GenerateCargo(Resource.Type type, float min, float max)
+        public Cargo GenerateCargo(Resource.Type type, uint min, uint max)
         {
             return new Cargo(type, min, max);
         }
